feat: use f64 for current for now.

+++ src/current/constant_current.rs
@@ -3,20 +3,20 @@ use crate::error::Error;
 use super::CurrentData;
 
 pub(crate) struct ConstantCurrent {
-    u: f32,
-    v: f32,
+    u: f64,
+    v: f64,
 }
 
 #[allow(dead_code)]
 impl ConstantCurrent {
     /// Constructor
-    pub(crate) fn new(u: f32, v: f32) -> Self {
+    pub(crate) fn new(u: f64, v: f64) -> Self {
         ConstantCurrent { u, v }
     }
 }
 
 impl CurrentData for ConstantCurrent {
-    fn current(&self, _x: &f32, _y: &f32) -> Result<(f32, f32), Error> {
+    fn current(&self, _x: &f64, _y: &f64) -> Result<(f64, f64), Error> {
         Ok((self.u, self.v))
     }
 }
